google_cloud_functions/image_preprocessing/requirements.txt
opencv-python-headless
google-cloud-storage
requests==2.26.0
google-auth
google-auth-oauthlib
numpy
google-cloud-firestore
